import: update Makefile for standalone use
The original Makefile was intended to run as part of the ATF build.
As we split this off into a standalone project again, we need to
change things and update/add rules that are missing when the ATF
build infrastructure is not present.

This commit makes a standalone Makefile from what we found within
the ATF repository.