src: rename main function to 'm0_main' to avoid warnings
With recent GCC versions (we currently use GCC-6.3.0 for code-size
and static analysis reasons), we can neither have a main()-function
that has a void return-type (warns on return not being 'int') nor
can we have a function with an 'int' return-type marked 'noreturn'.

Consequently we rename the main-function to 'm0_main' to help the
compiler understand that we are not in a standard C runtime.